backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.7.4
pydantic-settings>=2.1.0
email-validator>=2.1.0
qdrant-client>=1.7.1
python-dotenv>=1.0.0
asyncpg>=0.29.0
sqlalchemy>=2.0.25
bcrypt>=4.0.1
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
httpx>=0.26.0
google-generativeai>=0.8.0